requirements.txt
Flask==1.1.1
Flask-RESTX==0.1.1
werkzeug==0.16.1
python-dotenv==0.14.0
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.4
pytest==5.4.3
faker==4.1.1
gunicorn==20.0.4
pytest-cov==2.8.1
flake8==3.7.9